feat: add role-based access
TODO: Resource-ownership-based access

+++ models/motel.go
@@ -6,7 +6,7 @@ import (
 )
 
 type Motel struct {
-	MotelID       int    `json:"motel_id" db:"motel_id"`
+	MotelID       int64  `json:"motel_id" db:"motel_id"`
 	Name          string `json:"name" db:"name"`
 	Location      string `json:"location" db:"location"`
 	ContactNumber string `json:"contact_number" db:"contact_number"`
